requirements.txt
asgiref==3.5.2
Django==4.0.5
psycopg2-binary
sqlparse==0.4.2
pytest-django
pylint
pylint-django
djangorestframework
django-environ
factory_boy
Faker
requests
Celery
django-celery-results
django-celery-beat